extract ArrayMergeFromArraySpreadFactory
