+++ rules-tests/DowngradePhp74/Rector/Array_/DowngradeArraySpreadRector/Fixture/skip_property_default.php.inc
@@ -1,0 +1,18 @@
+<?php
+
+namespace Rector\Tests\DowngradePhp74\Rector\Array_\DowngradeArraySpreadRector\Fixture;
+
+final class SkipPropertyDefault
+{
+    /**
+     * @var string
+     */
+    private $items = [];
+
+    public function run()
+    {
+        if (in_array(1, $this->items)) {
+            return;
+        }
+    }
+}
