+++ rules-tests/DowngradePhp81/Rector/Array_/DowngradeArraySpreadStringKeyRector/Fixture/array_spread_string_key_by_doc.php.inc
@@ -2,16 +2,24 @@
 
 namespace Rector\Tests\DowngradePhp81\Rector\Array_\DowngradeArraySpreadStringKeyRector\Fixture;
 
-class ArraySpreadStringKeyByDoc
+final class ArraySpreadStringKeyByDoc
 {
     public function run()
     {
         /** @var array<string, string> $parts */
         $parts = $this->data();
         /** @var array<string, string> $parts2 */
-        $parts2 = $this->data2();
+        $parts2 = $this->data();
 
         $result = [...$parts, ...$parts2];
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    private function data(): array
+    {
+        return ['your' => 'name'];
     }
 }
 
@@ -21,16 +29,24 @@ class ArraySpreadStringKeyByDoc
 
 namespace Rector\Tests\DowngradePhp81\Rector\Array_\DowngradeArraySpreadStringKeyRector\Fixture;
 
-class ArraySpreadStringKeyByDoc
+final class ArraySpreadStringKeyByDoc
 {
     public function run()
     {
         /** @var array<string, string> $parts */
         $parts = $this->data();
         /** @var array<string, string> $parts2 */
-        $parts2 = $this->data2();
+        $parts2 = $this->data();
 
         $result = array_merge($parts, $parts2);
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    private function data(): array
+    {
+        return ['your' => 'name'];
     }
 }
