+++ tests/Issues/IssueDowngradeArraySpread/Fixture/skip_property_default.php.inc
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rector\Core\Tests\Issues\IssueDowngradeArraySpread\Fixture;
+
+final class SkipPropertyDefault
+{
+    /**
+     * @var string[]
+     */
+    private $defaultProperty = [];
+}
